chore(tests/account): moved tests file to match src fs tree
