+++ test/types/Account/_getBIP44Path.js
