+++ test/types/Account/_getNetwork.js
