+++ test/types/Account/methods/connect.js
