+++ test/types/Account/methods/fetchTransactionInfo.js
