+++ test/types/Account/methods/generateAddress.js
