+++ test/types/Account/methods/getAddresses.js
