+++ test/types/Account/methods/getPlugin.js
