+++ test/types/Account/methods/getUnusedAddress.js
@@ -1,11 +1,11 @@
 const { expect } = require('chai');
 const Dashcore = require('@dashevo/dashcore-lib');
-const getUnusedAddress = require('../../../src/types/Account/methods/getUnusedAddress');
-const getAddress = require('../../../src/types/Account/methods/getAddress');
-const generateAddress = require('../../../src/types/Account/methods/generateAddress');
-const KeyChain = require('../../../src/types/KeyChain/KeyChain');
+const getUnusedAddress = require('../../../../src/types/Account/methods/getUnusedAddress');
+const getAddress = require('../../../../src/types/Account/methods/getAddress');
+const generateAddress = require('../../../../src/types/Account/methods/generateAddress');
+const KeyChain = require('../../../../src/types/KeyChain/KeyChain');
 
-const mockedStore = require('../../fixtures/duringdevelop-fullstore-snapshot-1548538361');
+const mockedStore = require('../../../fixtures/duringdevelop-fullstore-snapshot-1548538361');
 
 const HDRootKeyMockedStore = 'tprv8ZgxMBicQKsPfEan1JB7NF4STbvnjGvP9318CN7FPGZp5nsUTBqmerxtDVpsJjFufyfkTgoe6QfHcDhMqjN3ZoFKtb8SnXFeubNjQreZSq6';
