+++ test/types/Account/methods/injectPlugin.js
