+++ test/types/Account/methods/sign.js
