+++ test/types/Account/methods/updateNetwork.js
